+++ configure.py
@@ -575,6 +575,7 @@ scylla_core = (['database.cc',
                 'sstables/mp_row_consumer.cc',
                 'sstables/sstables.cc',
                 'sstables/sstables_manager.cc',
+                'sstables/sstable_set.cc',
                 'sstables/mx/writer.cc',
                 'sstables/kl/writer.cc',
                 'sstables/sstable_version.cc',
