+++ sstables/compaction_strategy.cc
@@ -47,10 +47,8 @@
 #include "compaction_strategy_impl.hh"
 #include "schema.hh"
 #include "sstable_set.hh"
-#include "compatible_ring_position.hh"
 #include <boost/range/algorithm/find.hpp>
 #include <boost/range/adaptors.hpp>
-#include <boost/icl/interval_map.hpp>
 #include <boost/algorithm/cxx11/any_of.hpp>
 #include "size_tiered_compaction_strategy.hh"
 #include "date_tiered_compaction_strategy.hh"
@@ -63,384 +61,6 @@ logging::logger date_tiered_manifest::logger = logging::logger("DateTieredCompac
 logging::logger leveled_manifest::logger("LeveledManifest");
 
 namespace sstables {
-
-extern logging::logger clogger;
-
-void sstable_run::insert(shared_sstable sst) {
-    _all.insert(std::move(sst));
-}
-
-void sstable_run::erase(shared_sstable sst) {
-    _all.erase(sst);
-}
-
-uint64_t sstable_run::data_size() const {
-    return boost::accumulate(_all | boost::adaptors::transformed(std::mem_fn(&sstable::data_size)), uint64_t(0));
-}
-
-std::ostream& operator<<(std::ostream& os, const sstables::sstable_run& run) {
-    os << "Run = {\n";
-    if (run.all().empty()) {
-        os << "  Identifier: not found\n";
-    } else {
-        os << format("  Identifier: {}\n", (*run.all().begin())->run_identifier());
-    }
-
-    auto frags = boost::copy_range<std::vector<shared_sstable>>(run.all());
-    boost::sort(frags, [] (const shared_sstable& x, const shared_sstable& y) {
-        return x->get_first_decorated_key().token() < y->get_first_decorated_key().token();
-    });
-    os << "  Fragments = {\n";
-    for (auto& frag : frags) {
-        os << format("    {}={}:{}\n", frag->generation(), frag->get_first_decorated_key().token(), frag->get_last_decorated_key().token());
-    }
-    os << "  }\n}\n";
-    return os;
-}
-
-class incremental_selector_impl {
-public:
-    virtual ~incremental_selector_impl() {}
-    virtual std::tuple<dht::partition_range, std::vector<shared_sstable>, dht::ring_position_view> select(const dht::ring_position_view&) = 0;
-};
-
-class sstable_set_impl {
-public:
-    virtual ~sstable_set_impl() {}
-    virtual std::unique_ptr<sstable_set_impl> clone() const = 0;
-    virtual std::vector<shared_sstable> select(const dht::partition_range& range) const = 0;
-    virtual void insert(shared_sstable sst) = 0;
-    virtual void erase(shared_sstable sst) = 0;
-    virtual std::unique_ptr<incremental_selector_impl> make_incremental_selector() const = 0;
-};
-
-sstable_set::sstable_set(std::unique_ptr<sstable_set_impl> impl, schema_ptr s, lw_shared_ptr<sstable_list> all)
-        : _impl(std::move(impl))
-        , _schema(std::move(s))
-        , _all(std::move(all)) {
-}
-
-sstable_set::sstable_set(const sstable_set& x)
-        : _impl(x._impl->clone())
-        , _schema(x._schema)
-        , _all(make_lw_shared<sstable_list>(*x._all))
-        , _all_runs(x._all_runs) {
-}
-
-sstable_set::sstable_set(sstable_set&&) noexcept = default;
-
-sstable_set&
-sstable_set::operator=(const sstable_set& x) {
-    if (this != &x) {
-        auto tmp = sstable_set(x);
-        *this = std::move(tmp);
-    }
-    return *this;
-}
-
-sstable_set&
-sstable_set::operator=(sstable_set&&) noexcept = default;
-
-std::vector<shared_sstable>
-sstable_set::select(const dht::partition_range& range) const {
-    return _impl->select(range);
-}
-
-std::vector<sstable_run>
-sstable_set::select_sstable_runs(const std::vector<shared_sstable>& sstables) const {
-    auto run_ids = boost::copy_range<std::unordered_set<utils::UUID>>(sstables | boost::adaptors::transformed(std::mem_fn(&sstable::run_identifier)));
-    return boost::copy_range<std::vector<sstable_run>>(run_ids | boost::adaptors::transformed([this] (utils::UUID run_id) {
-        return _all_runs.at(run_id);
-    }));
-}
-
-void
-sstable_set::insert(shared_sstable sst) {
-    _impl->insert(sst);
-    try {
-        _all->insert(sst);
-        try {
-            _all_runs[sst->run_identifier()].insert(sst);
-        } catch (...) {
-            _all->erase(sst);
-            throw;
-        }
-    } catch (...) {
-        _impl->erase(sst);
-        throw;
-    }
-}
-
-void
-sstable_set::erase(shared_sstable sst) {
-    _impl->erase(sst);
-    _all->erase(sst);
-    _all_runs[sst->run_identifier()].erase(sst);
-}
-
-sstable_set::~sstable_set() = default;
-
-sstable_set::incremental_selector::incremental_selector(std::unique_ptr<incremental_selector_impl> impl, const schema& s)
-    : _impl(std::move(impl))
-    , _cmp(s) {
-}
-
-sstable_set::incremental_selector::~incremental_selector() = default;
-
-sstable_set::incremental_selector::incremental_selector(sstable_set::incremental_selector&&) noexcept = default;
-
-sstable_set::incremental_selector::selection
-sstable_set::incremental_selector::select(const dht::ring_position_view& pos) const {
-    if (!_current_range_view || !_current_range_view->contains(pos, _cmp)) {
-        std::tie(_current_range, _current_sstables, _current_next_position) = _impl->select(pos);
-        _current_range_view = _current_range->transform([] (const dht::ring_position& rp) { return dht::ring_position_view(rp); });
-    }
-    return {_current_sstables, _current_next_position};
-}
-
-sstable_set::incremental_selector
-sstable_set::make_incremental_selector() const {
-    return incremental_selector(_impl->make_incremental_selector(), *_schema);
-}
-
-// default sstable_set, not specialized for anything
-class bag_sstable_set : public sstable_set_impl {
-    // erasing is slow, but select() is fast
-    std::vector<shared_sstable> _sstables;
-public:
-    virtual std::unique_ptr<sstable_set_impl> clone() const override {
-        return std::make_unique<bag_sstable_set>(*this);
-    }
-    virtual std::vector<shared_sstable> select(const dht::partition_range& range = query::full_partition_range) const override {
-        return _sstables;
-    }
-    virtual void insert(shared_sstable sst) override {
-        _sstables.push_back(std::move(sst));
-    }
-    virtual void erase(shared_sstable sst) override {
-        auto it = boost::range::find(_sstables, sst);
-        if (it != _sstables.end()){
-            _sstables.erase(it);
-        }
-    }
-    virtual std::unique_ptr<incremental_selector_impl> make_incremental_selector() const override;
-    class incremental_selector;
-};
-
-class bag_sstable_set::incremental_selector : public incremental_selector_impl {
-    const std::vector<shared_sstable>& _sstables;
-public:
-    incremental_selector(const std::vector<shared_sstable>& sstables)
-        : _sstables(sstables) {
-    }
-    virtual std::tuple<dht::partition_range, std::vector<shared_sstable>, dht::ring_position_view> select(const dht::ring_position_view&) override {
-        return std::make_tuple(dht::partition_range::make_open_ended_both_sides(), _sstables, dht::ring_position_view::max());
-    }
-};
-
-std::unique_ptr<incremental_selector_impl> bag_sstable_set::make_incremental_selector() const {
-    return std::make_unique<incremental_selector>(_sstables);
-}
-
-// specialized when sstables are partitioned in the token range space
-// e.g. leveled compaction strategy
-class partitioned_sstable_set : public sstable_set_impl {
-    using value_set = std::unordered_set<shared_sstable>;
-    using interval_map_type = boost::icl::interval_map<compatible_ring_position_or_view, value_set>;
-    using interval_type = interval_map_type::interval_type;
-    using map_iterator = interval_map_type::const_iterator;
-private:
-    schema_ptr _schema;
-    std::vector<shared_sstable> _unleveled_sstables;
-    interval_map_type _leveled_sstables;
-    // Change counter on interval map for leveled sstables which is used by
-    // incremental selector to determine whether or not to invalidate iterators.
-    uint64_t _leveled_sstables_change_cnt = 0;
-    bool _use_level_metadata = false;
-private:
-    static interval_type make_interval(const schema& s, const dht::partition_range& range) {
-        return interval_type::closed(
-                compatible_ring_position_or_view(s, dht::ring_position_view(range.start()->value())),
-                compatible_ring_position_or_view(s, dht::ring_position_view(range.end()->value())));
-    }
-    interval_type make_interval(const dht::partition_range& range) const {
-        return make_interval(*_schema, range);
-    }
-    static interval_type make_interval(const schema_ptr& s, const sstable& sst) {
-        return interval_type::closed(
-                compatible_ring_position_or_view(s, dht::ring_position(sst.get_first_decorated_key())),
-                compatible_ring_position_or_view(s, dht::ring_position(sst.get_last_decorated_key())));
-    }
-    interval_type make_interval(const sstable& sst) {
-        return make_interval(_schema, sst);
-    }
-    interval_type singular(const dht::ring_position& rp) const {
-        // We should use the view here, since this is used for queries.
-        auto rpv = dht::ring_position_view(rp);
-        auto crp = compatible_ring_position_or_view(*_schema, std::move(rpv));
-        return interval_type::closed(crp, crp);
-    }
-    std::pair<map_iterator, map_iterator> query(const dht::partition_range& range) const {
-        if (range.start() && range.end()) {
-            return _leveled_sstables.equal_range(make_interval(range));
-        }
-        else if (range.start() && !range.end()) {
-            auto start = singular(range.start()->value());
-            return { _leveled_sstables.lower_bound(start), _leveled_sstables.end() };
-        } else if (!range.start() && range.end()) {
-            auto end = singular(range.end()->value());
-            return { _leveled_sstables.begin(), _leveled_sstables.upper_bound(end) };
-        } else {
-            return { _leveled_sstables.begin(), _leveled_sstables.end() };
-        }
-    }
-    // SSTables are stored separately to avoid interval map's fragmentation issue when level 0 falls behind.
-    bool store_as_unleveled(const shared_sstable& sst) const {
-        return _use_level_metadata && sst->get_sstable_level() == 0;
-    }
-public:
-    static dht::ring_position to_ring_position(const compatible_ring_position_or_view& crp) {
-        // Ring position views, representing bounds of sstable intervals are
-        // guaranteed to have key() != nullptr;
-        const auto& pos = crp.position();
-        return dht::ring_position(pos.token(), *pos.key());
-    }
-    static dht::partition_range to_partition_range(const interval_type& i) {
-        return dht::partition_range::make(
-                {to_ring_position(i.lower()), boost::icl::is_left_closed(i.bounds())},
-                {to_ring_position(i.upper()), boost::icl::is_right_closed(i.bounds())});
-    }
-    static dht::partition_range to_partition_range(const dht::ring_position_view& pos, const interval_type& i) {
-        auto lower_bound = [&] {
-            if (pos.key()) {
-                return dht::partition_range::bound(dht::ring_position(pos.token(), *pos.key()),
-                        pos.is_after_key() == dht::ring_position_view::after_key::no);
-            } else {
-                return dht::partition_range::bound(dht::ring_position(pos.token(), pos.get_token_bound()), true);
-            }
-        }();
-        auto upper_bound = dht::partition_range::bound(to_ring_position(i.lower()), !boost::icl::is_left_closed(i.bounds()));
-        return dht::partition_range::make(std::move(lower_bound), std::move(upper_bound));
-    }
-    explicit partitioned_sstable_set(schema_ptr schema, bool use_level_metadata = true)
-            : _schema(std::move(schema))
-            , _use_level_metadata(use_level_metadata) {
-    }
-    virtual std::unique_ptr<sstable_set_impl> clone() const override {
-        return std::make_unique<partitioned_sstable_set>(*this);
-    }
-    virtual std::vector<shared_sstable> select(const dht::partition_range& range) const override {
-        auto ipair = query(range);
-        auto b = std::move(ipair.first);
-        auto e = std::move(ipair.second);
-        value_set result;
-        while (b != e) {
-            boost::copy(b++->second, std::inserter(result, result.end()));
-        }
-        auto r = _unleveled_sstables;
-        r.insert(r.end(), result.begin(), result.end());
-        return r;
-    }
-    virtual void insert(shared_sstable sst) override {
-        if (store_as_unleveled(sst)) {
-            _unleveled_sstables.push_back(std::move(sst));
-        } else {
-            _leveled_sstables_change_cnt++;
-            _leveled_sstables.add({make_interval(*sst), value_set({sst})});
-        }
-    }
-    virtual void erase(shared_sstable sst) override {
-        if (store_as_unleveled(sst)) {
-            _unleveled_sstables.erase(std::remove(_unleveled_sstables.begin(), _unleveled_sstables.end(), sst), _unleveled_sstables.end());
-        } else {
-            _leveled_sstables_change_cnt++;
-            _leveled_sstables.subtract({make_interval(*sst), value_set({sst})});
-        }
-    }
-    virtual std::unique_ptr<incremental_selector_impl> make_incremental_selector() const override;
-    class incremental_selector;
-};
-
-class partitioned_sstable_set::incremental_selector : public incremental_selector_impl {
-    schema_ptr _schema;
-    const std::vector<shared_sstable>& _unleveled_sstables;
-    const interval_map_type& _leveled_sstables;
-    const uint64_t& _leveled_sstables_change_cnt;
-    uint64_t _last_known_leveled_sstables_change_cnt;
-    map_iterator _it;
-    // Only to back the dht::ring_position_view returned from select().
-    dht::ring_position _next_position;
-private:
-    dht::ring_position_view next_position(map_iterator it) {
-        if (it == _leveled_sstables.end()) {
-            _next_position = dht::ring_position::max();
-            return dht::ring_position_view::max();
-        } else {
-            _next_position = partitioned_sstable_set::to_ring_position(it->first.lower());
-            return dht::ring_position_view(_next_position, dht::ring_position_view::after_key(!boost::icl::is_left_closed(it->first.bounds())));
-        }
-    }
-    static bool is_before_interval(const compatible_ring_position_or_view& crp, const interval_type& interval) {
-        if (boost::icl::is_left_closed(interval.bounds())) {
-            return crp < interval.lower();
-        } else {
-            return crp <= interval.lower();
-        }
-    }
-    void maybe_invalidate_iterator(const compatible_ring_position_or_view& crp) {
-        if (_last_known_leveled_sstables_change_cnt != _leveled_sstables_change_cnt) {
-            _it = _leveled_sstables.lower_bound(interval_type::closed(crp, crp));
-            _last_known_leveled_sstables_change_cnt = _leveled_sstables_change_cnt;
-        }
-    }
-public:
-    incremental_selector(schema_ptr schema, const std::vector<shared_sstable>& unleveled_sstables, const interval_map_type& leveled_sstables,
-                         const uint64_t& leveled_sstables_change_cnt)
-        : _schema(std::move(schema))
-        , _unleveled_sstables(unleveled_sstables)
-        , _leveled_sstables(leveled_sstables)
-        , _leveled_sstables_change_cnt(leveled_sstables_change_cnt)
-        , _last_known_leveled_sstables_change_cnt(leveled_sstables_change_cnt)
-        , _it(leveled_sstables.begin())
-        , _next_position(dht::ring_position::min()) {
-    }
-    virtual std::tuple<dht::partition_range, std::vector<shared_sstable>, dht::ring_position_view> select(const dht::ring_position_view& pos) override {
-        auto crp = compatible_ring_position_or_view(*_schema, pos);
-        auto ssts = _unleveled_sstables;
-        using namespace dht;
-
-        maybe_invalidate_iterator(crp);
-
-        while (_it != _leveled_sstables.end()) {
-            if (boost::icl::contains(_it->first, crp)) {
-                ssts.insert(ssts.end(), _it->second.begin(), _it->second.end());
-                return std::make_tuple(partitioned_sstable_set::to_partition_range(_it->first), std::move(ssts), next_position(std::next(_it)));
-            }
-            // We don't want to skip current interval if pos lies before it.
-            if (is_before_interval(crp, _it->first)) {
-                return std::make_tuple(partitioned_sstable_set::to_partition_range(pos, _it->first), std::move(ssts), next_position(_it));
-            }
-            _it++;
-        }
-        return std::make_tuple(partition_range::make_open_ended_both_sides(), std::move(ssts), ring_position_view::max());
-    }
-};
-
-std::unique_ptr<incremental_selector_impl> partitioned_sstable_set::make_incremental_selector() const {
-    return std::make_unique<incremental_selector>(_schema, _unleveled_sstables, _leveled_sstables, _leveled_sstables_change_cnt);
-}
-
-std::unique_ptr<sstable_set_impl> compaction_strategy_impl::make_sstable_set(schema_ptr schema) const {
-    return std::make_unique<bag_sstable_set>();
-}
-
-std::unique_ptr<sstable_set_impl> leveled_compaction_strategy::make_sstable_set(schema_ptr schema) const {
-    return std::make_unique<partitioned_sstable_set>(std::move(schema));
-}
-
-sstable_set make_partitioned_sstable_set(schema_ptr schema, lw_shared_ptr<sstable_list> all, bool use_level_metadata) {
-    return sstables::sstable_set(std::make_unique<partitioned_sstable_set>(schema, use_level_metadata), schema, std::move(all));
-}
 
 compaction_descriptor compaction_strategy_impl::get_major_compaction_job(column_family& cf, std::vector<sstables::shared_sstable> candidates) {
     return compaction_descriptor(std::move(candidates), cf.get_sstable_set(), service::get_local_compaction_priority());
@@ -549,6 +169,8 @@ void size_tiered_backlog_tracker::remove_sstable(sstables::shared_sstable sst) {
 }
 
 namespace sstables {
+
+extern logging::logger clogger;
 
 // The backlog for TWCS is just the sum of the individual backlogs in each time window.
 // We'll keep various SizeTiered backlog tracker objects-- one per window for the static SSTables.
@@ -1029,14 +651,6 @@ int64_t compaction_strategy::estimated_pending_compactions(column_family& cf) co
 
 bool compaction_strategy::use_clustering_key_filter() const {
     return _compaction_strategy_impl->use_clustering_key_filter();
-}
-
-sstable_set
-compaction_strategy::make_sstable_set(schema_ptr schema) const {
-    return sstable_set(
-            _compaction_strategy_impl->make_sstable_set(schema),
-            schema,
-            make_lw_shared<sstable_list>());
 }
 
 compaction_backlog_tracker& compaction_strategy::get_backlog_tracker() {
